Basic support for arrays added.
In terms of the language, local/global variable declarations are extended to include arrays of any existing primitive type, and also multi-dimensional arrays:

* a : array[N] of [0..2];
* b : array[N] of bool;
* c : array[N] of array[M] of [0..2];

Array sizes must be constant expressions, and initial values can be provided in the usual style, but currently all array elements must be initialised to the same value:

* a : array[N] of [0..2] init 2;

Accesses to array variables are expressed using the notation a[i], where i is any integer-valued expression. Arrays are indexed starting from zero. Updates to array elements are also performed in the usual way, using primed notation:

* [] i<N -> (a[i]'=a[i]+1);

Support for model checking of models using arrays (and properties that refer to  them) has been added to the main engines, but it is not yet supported for PTAs. Several test cases can be found in the tests directory.

There is a small change to the parser required to enable this. Bounds attached to P/R/S operators may now need to be parenthesised if they are expressions, to avoid ambiguities in the grammar. Some common cases are still permitted, e.g. P<p[], P<1-p[], P<2*p[] are allowed, but P<(1-2*p)[] needs parentheses.

Code-level changes for supporting arrays include:

* Updates to PRISM language parsers to support arrays
* Arrays added to DeclarationType and Type class hierarchies
* New ExpressionArrayAccess AST element class
* VarList overhauled to maintain a list of “primitive” variables, and how all variables (including arrays) map to them.
* EvaluateContext changed from interface to abstract class; this base class stores a varIndexOffset, used to keep track of variable indexing computations when evaluating array-based expressions
* Variable indexing, used for evaluation of expressions, etc. updated and extended to support arrays. Info now stored in both ExpressionVar and ExpressionArrayAccess. This is now performed by the method VarList.addVarIndexing, and triggered each type constants are set on a model/properties file. The FindAllVars visitor just identifies variable references, but does no indexing (because constant values, now needed for array indexing, are not available at that point).
* Update: variable references now stored as Expressions to handle the fact that the updates may be to array elements.
* ModulesFile: Creates and caches a VarList each time that constant values are set (was previously created anew every time needed). Implementation of getDefaultInitialState() now pushed to VarList. It also remembers whether if we are in “exact” mode