Remove calls to evaluateExact and castFromBigRational methods.
These are generally replaced with a call to the new evaluate()
method, passing an EvaluateContext with the appropriate EvalMode.
Sometimes a call to castValueTo() is also used.

Relatedly, for convenience, prism.Evaluator now has an
evalMode() method returning the appropriate EvalMode.

All calls to castFromBigRational are gone and it is removed from Type.

The only remaining calls to evaluateExact are in the param engine.