Basic implementation of uncertain MDPs.
UMDPSimple class implements UMDP, with a similar storage to MDPSimple
but each state-choice is an instance of UDistribution. This allows any
uncertain distribution to be stored for each state-choice pair.

As a test case, UDistributionIntervals implements UDistribution,
and could in principle replace IMDPSimple, but is currently unused.

A first version of UDistributionL1Max is added but without
implementation of the inner optimisations mvMultUnc.